#45 [REFACTOR] 최근 검색창 페이지 관련 클래스 구조 변경
